+++ CHANGELOG.md
@@ -6,6 +6,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 For upcoming sub-apps and features see [Feature Tracker](https://github.com/Finoozer/PyPanel/projects/2)
 
+## [v0.1.6] - 2021-02-26
+### Added 
+* Icon to app executable
+* **About** window #1
+* Use entire URL or ID in RustApp #23 
+* Choose servers in **RustApp** from a list #30
+* GPG signature
+
+### Fixed
+* Table refreshing not working when deleting elements #4
+* User can rename `DEFAULT` profile #22
+* Default Profile Manager window size hides remove button #32 
+
+### Changed
+* Use entire URL or ID in RustApp #23
+* WeatherApp table should be without a scrollbar #28
+* One-file to One-dir app design #17
+  * Inno Setup Installer
+
 ## [v0.1.5] - 2021-02-22
 ### Added
 * Profile Manager